@@ -13,10 +13,10 @@
   ],
   "require": {
     "ext-json": "*",
-    "php": ">=7.2"
+    "php": ">=8.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "8.3.5"
+    "phpunit/phpunit": "~9.6"
   },
   "autoload": {
     "psr-4": {
