@@ -2,6 +2,7 @@
 
 namespace OpenRtb\BidRequest;
 
+use OpenRtb\Tools\Exceptions\ExceptionInvalidValue;
 use OpenRtb\Tools\Interfaces\Arrayable;
 use OpenRtb\BidRequest\Specification\BitType;
 use OpenRtb\BidRequest\Specification\ConnectionType;
@@ -250,7 +251,7 @@ class Device implements Arrayable
     /**
      * @param string $ua
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setUa($ua)
     {
@@ -288,7 +289,7 @@ class Device implements Arrayable
     /**
      * @param int $dnt
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setDnt($dnt)
     {
@@ -308,7 +309,7 @@ class Device implements Arrayable
     /**
      * @param int $lmt
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setLmt($lmt)
     {
@@ -328,7 +329,7 @@ class Device implements Arrayable
     /**
      * @param string $ip
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setIp($ip)
     {
@@ -348,7 +349,7 @@ class Device implements Arrayable
     /**
      * @param string $ipv6
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setIpv6($ipv6)
     {
@@ -368,7 +369,7 @@ class Device implements Arrayable
     /**
      * @param int $devicetype
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setDevicetype($devicetype)
     {
@@ -388,7 +389,7 @@ class Device implements Arrayable
     /**
      * @param string $make
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setMake($make)
     {
@@ -408,7 +409,7 @@ class Device implements Arrayable
     /**
      * @param string $model
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setModel($model)
     {
@@ -428,7 +429,7 @@ class Device implements Arrayable
     /**
      * @param string $os
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setOs($os)
     {
@@ -448,7 +449,7 @@ class Device implements Arrayable
     /**
      * @param string $osv
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setOsv($osv)
     {
@@ -468,7 +469,7 @@ class Device implements Arrayable
     /**
      * @param string $hwv
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setHwv($hwv)
     {
@@ -488,7 +489,7 @@ class Device implements Arrayable
     /**
      * @param int $h
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setH($h)
     {
@@ -507,7 +508,7 @@ class Device implements Arrayable
     /**
      * @param int $w
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setW($w)
     {
@@ -526,7 +527,7 @@ class Device implements Arrayable
     /**
      * @param int $ppi
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setPpi($ppi)
     {
@@ -545,7 +546,7 @@ class Device implements Arrayable
     /**
      * @param float $pxratio
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setPxratio($pxratio)
     {
@@ -564,7 +565,7 @@ class Device implements Arrayable
     /**
      * @param int $js
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setJs($js)
     {
@@ -584,7 +585,7 @@ class Device implements Arrayable
     /**
      * @param string $flashver
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setFlashver($flashver)
     {
@@ -604,7 +605,7 @@ class Device implements Arrayable
     /**
      * @param string $language
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setLanguage($language)
     {
@@ -624,7 +625,7 @@ class Device implements Arrayable
     /**
      * @param string $carrier
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setCarrier($carrier)
     {
@@ -644,7 +645,7 @@ class Device implements Arrayable
     /**
      * @param int $connectiontype
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setConnectiontype($connectiontype)
     {
@@ -664,7 +665,7 @@ class Device implements Arrayable
     /**
      * @param string $ifa
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setIfa($ifa)
     {
@@ -684,7 +685,7 @@ class Device implements Arrayable
     /**
      * @param string $didsha1
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setDidsha1($didsha1)
     {
@@ -704,7 +705,7 @@ class Device implements Arrayable
     /**
      * @param string $didmd5
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setDidmd5($didmd5)
     {
@@ -724,7 +725,7 @@ class Device implements Arrayable
     /**
      * @param string $dpidsha1
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setDpidsha1($dpidsha1)
     {
@@ -744,7 +745,7 @@ class Device implements Arrayable
     /**
      * @param string $dpidmd5
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setDpidmd5($dpidmd5)
     {
@@ -764,7 +765,7 @@ class Device implements Arrayable
     /**
      * @param string $macsha1
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setMacsha1($macsha1)
     {
@@ -784,7 +785,7 @@ class Device implements Arrayable
     /**
      * @param string $macmd5
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setMacmd5($macmd5)
     {
@@ -822,7 +823,7 @@ class Device implements Arrayable
     /**
      * @param $geofetch
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setGeofetch($geofetch)
     {
@@ -843,7 +844,7 @@ class Device implements Arrayable
     /**
      * @param $mccmnc
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setMccmnc($mccmnc)
     {
@@ -852,6 +853,4 @@ class Device implements Arrayable
 
         return $this;
     }
-
-
 }
