@@ -3,6 +3,7 @@
 namespace OpenRtb\BidResponse;
 
 use OpenRtb\BidResponse\Specification\BitType;
+use OpenRtb\Tools\Exceptions\ExceptionInvalidValue;
 use OpenRtb\Tools\Interfaces\Arrayable;
 use OpenRtb\Tools\Traits\SetterValidation;
 use OpenRtb\Tools\Traits\ToArray;
@@ -91,7 +92,7 @@ class Seatbid implements Arrayable
     /**
      * @param string $seat
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setSeat($seat)
     {
@@ -111,7 +112,7 @@ class Seatbid implements Arrayable
     /**
      * @param int $group
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setGroup($group)
     {
