@@ -3,6 +3,7 @@
 namespace OpenRtb\NativeAdRequest;
 
 use OpenRtb\NativeAdRequest\Specification\BitType;
+use OpenRtb\Tools\Exceptions\ExceptionInvalidValue;
 use OpenRtb\Tools\Interfaces\Arrayable;
 use OpenRtb\Tools\Traits\SetterValidation;
 use OpenRtb\Tools\Traits\ToArray;
@@ -62,7 +63,7 @@ class Assets implements Arrayable
     /**
      * @param int $id
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setId($id)
     {
@@ -81,7 +82,7 @@ class Assets implements Arrayable
     /**
      * @param int $required
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setRequired($required)
     {
