@@ -3,6 +3,7 @@
 namespace OpenRtb\NativeAdRequest;
 
 use OpenRtb\NativeAdRequest\Specification\DataAssetType;
+use OpenRtb\Tools\Exceptions\ExceptionInvalidValue;
 use OpenRtb\Tools\Interfaces\Arrayable;
 use OpenRtb\Tools\Traits\SetterValidation;
 use OpenRtb\Tools\Traits\ToArray;
@@ -41,7 +42,7 @@ class Data implements Arrayable
     /**
      * @param int $type
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setType($type)
     {
@@ -61,7 +62,7 @@ class Data implements Arrayable
     /**
      * @param int $len
      * @return $this
-     * @throws \OpenRtb\Tools\Exceptions\ExceptionInvalidValue
+     * @throws ExceptionInvalidValue
      */
     public function setLen($len)
     {
