@@ -79,17 +79,13 @@ class ParametersBag implements Bag
     /**
      * @return ArrayIterator
      */
-    public function getIterator(): \Traversable
+    public function getIterator(): ArrayIterator
     {
         return new ArrayIterator($this->parameters);
     }
 
-    /**
-     * @return int
-     */
     public function count(): int
     {
         return count($this->parameters);
     }
-
 }
